+++ game/src/video/inc/video/TextureManager.h
@@ -47,7 +47,7 @@ public:
 	float GetFrameWidth(size_t texIndex, size_t /*frameIdx*/) const { return _logicalTextures[texIndex].pxFrameWidth; }
 	float GetFrameHeight(size_t texIndex, size_t /*frameIdx*/) const { return _logicalTextures[texIndex].pxFrameHeight; }
 	float GetBorderSize(size_t texIndex) const { return _logicalTextures[texIndex].pxBorderSize; }
-	unsigned int GetFrameCount(size_t texIndex) const { return _logicalTextures[texIndex].uvFrames.size(); }
+	unsigned int GetFrameCount(size_t texIndex) const { return static_cast<unsigned int>(_logicalTextures[texIndex].uvFrames.size()); }
 
 	void GetTextureNames(std::vector<std::string> &names, const char *prefix, bool noPrefixReturn) const;
